project: Upgrade to C++ 20 and MinGW build fix

+++ src/emu/disasm/src/disasm.cpp
@@ -61,7 +61,7 @@ namespace eka2l1 {
         cs_err err = cs_option(cp_handle, CS_OPT_MODE, thumb ? CS_MODE_THUMB : CS_MODE_ARM);
 
         if (err != CS_ERR_OK) {
-            LOG_ERROR(DISASM, "Unable to set disassemble option! Error: {}", err);
+            LOG_ERROR(DISASM, "Unable to set disassemble option! Error: {}", static_cast<int>(err));
             return "";
         }
